Fix Appevyor builds for non-release jobs
AppVeyor does not recommend providing unencrypted environment variables
to non-origin repos. Consequently, the PR builds trying to access the
DOCKER_USERNAME and DOCKER_PASSWORD fail because they are only
accessible to builds from the origin repo. Removed the docker login
command on PR builds and left it for release builds.